No need to recreate stack map frames, because CFG does not join
We do not change the control-flow graph of a method in such a way that
it becomes a cyclic graph when it was not already one.
See:
http://stackoverflow.com/questions/25109942/is-there-a-better-explanation-of-stack-map-frames

Also, we need not recompute the maxLocals, since it is just original
maxLocals + 1. The maxStack does not change, because no code we inject
will require a deeper stack than what was already needed for the
original method code.