Only create a TryCatchBlockSorter when method catches
This is a huge performance improvement, because the TryCatchBlockSorter
internally builds a complete instruction list of the method and emits
the ASM visit invocations at visitEnd().
And we only need that TryCatchBlockSorter when we actually have to
(re-)sort the try-catch blocks. And this is only when the original
method catches itself.