Transform private methods to take the MemoryStack as parameter
and patch all callsites which make use of the stack to delegate that
stack to the transformed method.
Also retain the original method for cases where we could not reliably
identify the call or where the caller does not use any stack methods
itself. This is the case for synthetic bridge methods for inner and
anonymous classes and for reflective invocations.